Rename .adb to .ato
- Update docs description of the data provider.
- Add "Changes" section to doc/ato.rst to record changes.
- Update base URL for remote files.
- Update org ID and contacts.
- Update tests.
