+++ transport_data/ato/__init__.py
@@ -1,4 +1,4 @@
-"""Asian Development Bank (ADB) provider."""
+"""Asian Transport Observatory (ATO) provider."""
 
 from collections import defaultdict
 from itertools import chain, product
@@ -13,7 +13,7 @@ from transport_data.util.pluggy import hookimpl
 from transport_data.util.pooch import Pooch
 from transport_data.util.sdmx import anno_generated
 
-BASE_URL = "https://asiantransportoutlook.com/exportdl?orig=1"
+BASE_URL = "https://asiantransportobservatory.org/exportdl?orig=1"
 
 #: List of all "ECONOMY" codes appearing in processed data.
 CL_ECONOMY = m.Codelist(
@@ -30,7 +30,7 @@ CS_MEASURE = m.ConceptScheme(
     id="MEASURE",
     name="Asian Transport Outlook measures (indicators)",
     description="Item are automatically generated by the TDCI tools. Currently no "
-    "correspondence with the list provided directly by ADB is checked or enforced.",
+    "correspondence with the list provided directly by ATO is checked or enforced.",
 )
 
 #: Mapping from short codes for ATO data categories to file names.
@@ -216,7 +216,7 @@ def dataset_to_metadata_reports(
         mdr.metadata.append(make_ra(mda_id, value))
 
     # Retrieve the MEASURE from CS_MEASURE
-    cs_measure = STORE.get("ConceptScheme=ADB:MEASURE(0.1.0)")
+    cs_measure = STORE.get("ConceptScheme=ATO:MEASURE(0.1.0)")
     measure_concept = cs_measure[dsd.id]
     # Construct a metadata attribute
     mdr.metadata.append(make_ra("MEASURE", measure_concept.name))
@@ -310,15 +310,18 @@ def format_data_provider(value: str) -> str:
 @hookimpl
 def get_agencies():
     a = m.Agency(
-        id="ADB",
-        name="Asian Transport Outlook team at the Asian Development Bank",
-        description="""See https://www.adb.org/what-we-do/topics/transport/asian-transport-outlook""",  # noqa: E501
+        id="ATO",
+        name="Asian Transport Observatory",
+        description="""See https://asiantransportobservatory.org/aboutus/""",  # noqa: E501
+        contact=[
+            m.Contact(email=["info@asiantransportobservatory.org"]),
+            m.Contact(name="Sudhir Gota", email=["sudhirgota@gmail.com"]),
+            # NB Affiliated with ADB, which supports ATO
+            # m.Contact(name="James Leather", email=["jleather@adb.org"]),
+            # NB No longer formally affiliated with ATO
+            # m.Contact(name="Cornie Huizenga", email=["chuizenga@cesg.biz"]),
+        ],
     )
-
-    c1 = m.Contact(name="James Leather", email=["jleather@adb.org"])
-    c2 = m.Contact(name="Cornie Huizenga", email=["chuizenga@cesg.biz"])
-    c3 = m.Contact(name="Sudhir Gota", email=["sudhirgota@gmail.com"])
-    a.contact.extend([c1, c2, c3])
 
     return (a,)
 
@@ -349,7 +352,7 @@ def prepare(aa: m.AnnotableArtefact) -> Tuple[m.DataSet, Callable]:
     CS_MEASURE.append(c)
 
     # Data structure definition with an ID matching the measure
-    # NB here we set ADB as the maintainer. Precisely, ADB establishes the data
+    # NB here we set ATO as the maintainer. Precisely, ATO establishes the data
     #    structure, but TDCI is maintaining the SDMX representation of it.
     ma_args = dict(maintainer=get_agencies()[0], version=VERSION)
     dsd = m.DataStructureDefinition(id=measure_id, **ma_args)
