+++ transport_data/proto/__init__.py
@@ -2,8 +2,8 @@
 
 Use::
 
-    $ tdc adb fetch --all --go
-    $ tdc adb convert TAS
+    $ tdc ato fetch --all --go
+    $ tdc ato convert TAS
     $ tdc jrc fetch --all --go
     $ tdc jrc convert --all
 """
